\name{decode_cstring}
\alias{decode_cstring}
\title{Deserialize null elements...}
\usage{decode_cstring(raw)}
\description{Deserialize null elements}
\details{The natural R type to the BSON Null value is NULL.}
\value{a named list whose single element is a char}
\arguments{\item{raw}{a raw vector}}
