\name{encode_int32}
\alias{encode_int32}
\title{Functions for BSON int32 type...}
\usage{encode_int32(num)}
\description{Functions for BSON int32 type}
\details{The BSON int32 corresponds to the R numeric type.}
\arguments{\item{num}{a R numeric to convert}
\item{raw}{a raw vector to convert}
\item{name}{the name of a int32 BSON element}}
