\name{encode_int64}
\alias{encode_int64}
\title{Functions for BSON int64 type...}
\usage{encode_int64(num)}
\description{Functions for BSON int64 type}
\details{The BSON int64 corresponds to the R numeric type.}
\arguments{\item{num}{a R numeric to convert}
\item{raw}{a raw vector to convert}
\item{name}{the name of a int32 BSON element}}
