\name{encode_logical}
\alias{encode_logical}
\title{Functions for BSON boolean type...}
\usage{encode_logical(bool)}
\description{Functions for BSON boolean type}
\details{The BSON boolean corresponds to the R numeric type.}
\arguments{\item{num}{a R boolean to convert}
\item{raw}{a raw vector to convert}
\item{name}{the name of a boolean BSON element}}
